test: consolidate SQL Issue#NNNN test classes into thematic suites
Move @Test methods from 11 Issue*Test classes in
engine/src/test/java/com/arcadedb/query/sql/ into existing thematic
classes. Each migrated method carries a single-line // Issue #NNNN
comment for regression traceability.

Migrations:
- QueryTest (+40): Issue1625 (bucket params), Issue1898 (SELECT OR),
  Issue1972 (LIMIT/SKIP w/ params and LET), Issue3516 (named param
  offset), Issue3571 (@rid/@type/@cat on function results), Issue3581
  (CONTAINSANY method LHS), Issue3583 (replace().ilike() in LET),
  IssueContainsAllRidString, IssueLinkDotNotationProjection
- ExpandParseTest (+5): Issue2965 (expand + nested projection)
- DDLTest (+2): Issue4227 (CREATE INDEX missing type parsing)

Type/bucket names suffixed where needed to avoid collisions under the
shared TestHelper per-class database lifecycle.

Verified: engine test-compile clean; 76 tests pass across the three
modified classes (QueryTest 66, ExpandParseTest 6, DDLTest 4).
